requirement.txt
aiofiles==24.1.0
aiosqlite==0.20.0
asyncua==1.1.5
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
paho-mqtt==2.1.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sortedcontainers==2.4.0
typing_extensions==4.12.2
Werkzeug==3.1.3
